Clean-up docs, better destroying
