+++ packages/text-bitmap/src/utils/drawGlyph.ts
@@ -1,0 +1,90 @@
+import { generateFillStyle } from './generateFillStyle';
+import { hex2rgb, string2hex } from '@pixi/utils';
+import type { TextMetrics, TextStyle } from '@pixi/text';
+
+// TODO: Prevent code duplication b/w drawGlyph & Text#updateText
+
+/**
+ * Draws the glyph `metrics.text` on the given canvas.
+ *
+ * Ignored because not directly exposed.
+ *
+ * @ignore
+ * @param {HTMLCanvasElement} canvas
+ * @param {CanvasRenderingContext2D} context
+ * @param {TextMetrics} metrics
+ * @param {number} x
+ * @param {number} y
+ * @param {number} resolution
+ * @param {TextStyle} style
+ */
+export function drawGlyph(
+    canvas: HTMLCanvasElement,
+    context: CanvasRenderingContext2D,
+    metrics: TextMetrics,
+    x: number,
+    y: number,
+    resolution: number,
+    style: TextStyle
+): void
+{
+    const char = metrics.text;
+    const fontProperties = metrics.fontProperties;
+
+    context.translate(x, y);
+    context.scale(resolution, resolution);
+
+    const tx = style.strokeThickness / 2;
+    const ty = -(style.strokeThickness / 2);
+
+    context.font = style.toFontString();
+    context.lineWidth = style.strokeThickness;
+    context.textBaseline = style.textBaseline;
+    context.lineJoin = style.lineJoin;
+    context.miterLimit = style.miterLimit;
+
+    // set canvas text styles
+    context.fillStyle = generateFillStyle(canvas, context, style, resolution, [char], metrics);
+    context.strokeStyle = style.stroke as string;
+
+    context.font = style.toFontString();
+    context.lineWidth = style.strokeThickness;
+    context.textBaseline = style.textBaseline;
+    context.lineJoin = style.lineJoin;
+    context.miterLimit = style.miterLimit;
+
+    // set canvas text styles
+    context.fillStyle = generateFillStyle(canvas, context, style, resolution, [char], metrics);
+    context.strokeStyle = style.stroke as string;
+
+    const dropShadowColor = style.dropShadowColor;
+    const rgb = hex2rgb(typeof dropShadowColor === 'number' ? dropShadowColor : string2hex(dropShadowColor));
+
+    if (style.dropShadow)
+    {
+        context.shadowColor = `rgba(${rgb[0] * 255},${rgb[1] * 255},${rgb[2] * 255},${style.dropShadowAlpha})`;
+        context.shadowBlur = style.dropShadowBlur;
+        context.shadowOffsetX = Math.cos(style.dropShadowAngle) * style.dropShadowDistance;
+        context.shadowOffsetY = Math.sin(style.dropShadowAngle) * style.dropShadowDistance;
+    }
+    else
+    {
+        context.shadowColor = '0';
+        context.shadowBlur = 0;
+        context.shadowOffsetX = 0;
+        context.shadowOffsetY = 0;
+    }
+
+    if (style.stroke && style.strokeThickness)
+    {
+        context.strokeText(char, tx, ty + metrics.lineHeight - fontProperties.descent);
+    }
+    if (style.fill)
+    {
+        context.fillText(char, tx, ty + metrics.lineHeight - fontProperties.descent);
+    }
+
+    context.setTransform();
+
+    context.fillStyle = 'rgba(0, 0, 0, 0)';
+}
