+++ packages/text-bitmap/src/utils/index.ts
@@ -1,0 +1,3 @@
+export * from './drawGlyph';
+export * from './generateFillStyle';
+export * from './resolveCharacters';
